Implemented user paying backend
Implemented in UserBookingController the payment functionality.

When user requests booking, a Payment record is created at the same time for the same booking ID. Status is pending by default of the payment, and later when the user pays using the other route method the transactionID is set if non-cash payment, status is changed to completed, and paidAt is set to the current date.

No routes yet for this functionality, I will still test it and fix bugs as needed. Should work as expected though